Extend better AddOrRemoveAuthorInputFields
