+++ layers/Schema/packages/users/src/SchemaHooks/AbstractAddAuthorInputFieldsInputObjectTypeHookSet.php
@@ -1,0 +1,139 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\Users\SchemaHooks;
+
+use PoP\ComponentModel\Schema\SchemaTypeModifiers;
+use PoP\ComponentModel\TypeResolvers\InputObjectType\HookNames;
+use PoP\ComponentModel\TypeResolvers\InputObjectType\InputObjectTypeResolverInterface;
+use PoP\ComponentModel\TypeResolvers\InputTypeResolverInterface;
+use PoP\Engine\TypeResolvers\ScalarType\IDScalarTypeResolver;
+use PoP\Engine\TypeResolvers\ScalarType\StringScalarTypeResolver;
+use PoP\Hooks\AbstractHookSet;
+use PoPSchema\Users\ConditionalOnComponent\CustomPosts\FilterInputProcessors\FilterInputProcessor;
+
+abstract class AbstractAddAuthorInputFieldsInputObjectTypeHookSet extends AbstractHookSet
+{
+    use AddOrRemoveAuthorInputFieldsInputObjectTypeHookSetTrait;
+
+    private ?IDScalarTypeResolver $idScalarTypeResolver = null;
+    private ?StringScalarTypeResolver $stringScalarTypeResolver = null;
+
+    final public function setIDScalarTypeResolver(IDScalarTypeResolver $idScalarTypeResolver): void
+    {
+        $this->idScalarTypeResolver = $idScalarTypeResolver;
+    }
+    final protected function getIDScalarTypeResolver(): IDScalarTypeResolver
+    {
+        return $this->idScalarTypeResolver ??= $this->instanceManager->getInstance(IDScalarTypeResolver::class);
+    }
+    final public function setStringScalarTypeResolver(StringScalarTypeResolver $stringScalarTypeResolver): void
+    {
+        $this->stringScalarTypeResolver = $stringScalarTypeResolver;
+    }
+    final protected function getStringScalarTypeResolver(): StringScalarTypeResolver
+    {
+        return $this->stringScalarTypeResolver ??= $this->instanceManager->getInstance(StringScalarTypeResolver::class);
+    }
+
+    protected function init(): void
+    {
+        $this->getHooksAPI()->addFilter(
+            HookNames::INPUT_FIELD_NAME_TYPE_RESOLVERS,
+            [$this, 'getInputFieldNameTypeResolvers'],
+            10,
+            2
+        );
+        $this->getHooksAPI()->addFilter(
+            HookNames::INPUT_FIELD_DESCRIPTION,
+            [$this, 'getInputFieldDescription'],
+            10,
+            3
+        );
+        $this->getHooksAPI()->addFilter(
+            HookNames::INPUT_FIELD_TYPE_MODIFIERS,
+            [$this, 'getInputFieldTypeModifiers'],
+            10,
+            3
+        );
+        $this->getHooksAPI()->addFilter(
+            HookNames::INPUT_FIELD_FILTER_INPUT,
+            [$this, 'getInputFieldFilterInput'],
+            10,
+            3
+        );
+    }
+
+    /**
+     * Indicate if to add the fields added by the SchemaHookSet
+     */
+    abstract protected function addAuthorInputFields(
+        InputObjectTypeResolverInterface $inputObjectTypeResolver,
+    ): bool;
+
+    /**
+     * @param array<string, InputTypeResolverInterface> $inputFieldNameTypeResolvers
+     */
+    public function getInputFieldNameTypeResolvers(
+        array $inputFieldNameTypeResolvers,
+        InputObjectTypeResolverInterface $inputObjectTypeResolver,
+    ): array {
+        if (!$this->addAuthorInputFields($inputObjectTypeResolver)) {
+            return $inputFieldNameTypeResolvers;
+        }
+        return array_merge(
+            $inputFieldNameTypeResolvers,
+            $this->getAuthorInputFieldNameTypeResolvers(),
+        );
+    }
+
+    public function getInputFieldDescription(
+        ?string $inputFieldDescription,
+        InputObjectTypeResolverInterface $inputObjectTypeResolver,
+        string $inputFieldName
+    ): ?string {
+        if (!$this->addAuthorInputFields($inputObjectTypeResolver)) {
+            return $inputFieldDescription;
+        }
+        return match ($inputFieldName) {
+            'authorIDs' => $this->getTranslationAPI()->__('Get results from the authors with given IDs', 'pop-users'),
+            'authorSlug' => $this->getTranslationAPI()->__('Get results from the authors with given slug', 'pop-users'),
+            'excludeAuthorIDs' => $this->getTranslationAPI()->__('Get results excluding the ones from authors with given IDs', 'pop-users'),
+            default => $inputFieldDescription,
+        };
+    }
+
+    public function getInputFieldTypeModifiers(
+        int $inputFieldTypeModifiers,
+        InputObjectTypeResolverInterface $inputObjectTypeResolver,
+        string $inputFieldName
+    ): int {
+        if (!$this->addAuthorInputFields($inputObjectTypeResolver)) {
+            return $inputFieldTypeModifiers;
+        }
+        return match ($inputFieldName) {
+            'authorIDs',
+            'excludeAuthorIDs'
+                => SchemaTypeModifiers::IS_ARRAY | SchemaTypeModifiers::IS_NON_NULLABLE_ITEMS_IN_ARRAY,
+            default
+                => $inputFieldTypeModifiers,
+        };
+    }
+
+    public function getInputFieldFilterInput(
+        ?array $inputFieldFilterInput,
+        InputObjectTypeResolverInterface $inputObjectTypeResolver,
+        string $inputFieldName,
+    ): ?array {
+        if (!$this->addAuthorInputFields($inputObjectTypeResolver)) {
+            return $inputFieldFilterInput;
+        }
+        return match ($inputFieldName) {
+            'authorIDs' => [FilterInputProcessor::class, FilterInputProcessor::FILTERINPUT_AUTHOR_IDS],
+            'authorSlug' => [FilterInputProcessor::class, FilterInputProcessor::FILTERINPUT_AUTHOR_SLUG],
+            'excludeAuthorIDs' => [FilterInputProcessor::class, FilterInputProcessor::FILTERINPUT_EXCLUDE_AUTHOR_IDS],
+            default => $inputFieldFilterInput,
+        };
+    }
+}
